+++ lib/input_validation/ValidationHelpers.ts
@@ -15,6 +15,7 @@ import { IExpectedVerifiablePresentation, IExpectedVerifiableCredential, IExpect
 import LinkedDataCryptoSuitePublicKey from './LinkedDataCryptoSuitePublicKey';
 import ErrorHelpers from '../error_handling/ErrorHelpers';
 import ValidationError from '../error_handling/ValidationError';
+import { AuthenticationErrorCode } from '../error_handling/AuthenticationErrorCode';
 const errorCode = (error: number) => ErrorHelpers.errorCode('VCSDKVaHe', error);
 
 require('es6-promise').polyfill();
@@ -71,7 +72,8 @@ export class ValidationHelpers {
           code: errorCode(1),
           detailedError: `The ${(self as ValidationOptions).tokenType} could not be deserialized`,
           innerError: exception,
-          status: 400
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
         };
       }
     }
@@ -81,7 +83,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(2),
         detailedError: `The signature in the ${(self as ValidationOptions).tokenType} has an invalid format`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
       };
     }
 
@@ -92,7 +95,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(3),
           detailedError: `The payload in the ${(self as ValidationOptions).tokenType} is undefined`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
         };
       }
 
@@ -105,7 +109,8 @@ export class ValidationHelpers {
           code: errorCode(4),
           innerError: err,
           detailedError: `The payload in the ${(self as ValidationOptions).tokenType} is no valid JSON`,
-          status: 400
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
         };
       }
 
@@ -115,7 +120,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(5),
           detailedError: `The protected header in the ${(self as ValidationOptions).tokenType} does not contain the kid`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
         };
       }
 
@@ -136,7 +142,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(6),
           detailedError: `The proof is not available in the json ld payload`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
         };
       }
       if (!proof.verificationMethod) {
@@ -144,7 +151,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(7),
           detailedError: `The proof does not contain the verificationMethod in the json ld payload`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
         };
       }
       validationResponse.didKid = proof.verificationMethod;
@@ -172,7 +180,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(8),
           detailedError: `Could not retrieve DID document '${validationResponse.did}'`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         }
       }
       validationResponse.didDocument = resolveResult.didDocument;
@@ -184,7 +193,8 @@ export class ValidationHelpers {
         code: errorCode(9),
         innerError: err,
         detailedError: `Could not resolve DID '${validationResponse.did}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -194,7 +204,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(10),
         detailedError: `The kid is not referenced in the request`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
       };
     }
 
@@ -207,7 +218,8 @@ export class ValidationHelpers {
         code: errorCode(11),
         detailedError: exception.message,
         innerError: exception,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -225,11 +237,11 @@ export class ValidationHelpers {
     if (publicKey) {
       // Old ion protocol, to be deleted after switch
       signingKey = LinkedDataCryptoSuitePublicKey.getPublicKey(publicKey);
-    } else { 
+    } else {
       if (didDocument.rawDocument?.verificationMethod) {
         const keyIdParts = kid.split('#');
         const keyId = keyIdParts[keyIdParts.length - 1];
-        const verification = didDocument.rawDocument?.verificationMethod.filter ((vm: any) => vm.id.includes(`#${keyId}`));
+        const verification = didDocument.rawDocument?.verificationMethod.filter((vm: any) => vm.id.includes(`#${keyId}`));
         if (verification) {
           signingKey = LinkedDataCryptoSuitePublicKey.getPublicKey(verification[0]);
         }
@@ -267,7 +279,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(12),
           detailedError: `The presented ${(self as ValidationOptions).tokenType} is expired ${exp}, now ${current as number}`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
     }
@@ -283,7 +296,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(40),
           detailedError: `The presented ${(self as ValidationOptions).tokenType} is not yet valid ${nbf}`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
     }
@@ -306,7 +320,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(13),
         detailedError: `The issuer in configuration was not found`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -315,7 +330,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(14),
         detailedError: `Missing iss property in idToken. Expected '${JSON.stringify(issuer)}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -324,7 +340,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(15),
         detailedError: `The issuer in configuration '${issuer}' does not correspond with the issuer in the payload ${validationResponse.issuer}`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -332,9 +349,10 @@ export class ValidationHelpers {
     if (expected.audience && expected.audience !== validationResponse.payloadObject.aud) {
       return {
         result: false,
-        status: 401,
         code: errorCode(16),
-        detailedError: `The audience ${validationResponse.payloadObject.aud} is invalid`
+        detailedError: `The audience ${validationResponse.payloadObject.aud} is invalid`,
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -357,7 +375,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(17),
         detailedError: `Missing iss property in verifiablePresentation. Expected '${siopDid}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -366,7 +385,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(18),
         detailedError: `Wrong iss property in verifiablePresentation. Expected '${siopDid}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -377,7 +397,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(19),
           detailedError: `Missing aud property in verifiablePresentation. Expected '${expected.didAudience}'`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
 
@@ -386,7 +407,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(20),
           detailedError: `Wrong aud property in verifiablePresentation. Expected '${expected.didAudience}'. Found '${validationResponse.payloadObject.aud}'`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
     }
@@ -410,7 +432,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(21),
         detailedError: `Missing sub property in verifiableCredential. Expected '${siopDid}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -420,7 +443,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(22),
         detailedError: `Wrong sub property in verifiableCredential. Expected '${siopDid}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -436,24 +460,14 @@ export class ValidationHelpers {
   public checkScopeValidityOnSiopToken(validationResponse: IValidationResponse, expected: IExpectedSiop): IValidationResponse {
     const self: any = this;
 
-    // check sub
-    /* TODO temporary disabled
-    if (validationResponse.payloadObject.sub && validationResponse.payloadObject.sub !== validationResponse.did) {
-      return {
-        result: false,
-        detailedError: `The sub property in the siopIssuance must be equal to ${validationResponse.did}`,
-        status: 403
-      };
-    }
-    */
-
     // check iss value
     if (!validationResponse.issuer) {
       return validationResponse = {
         result: false,
         code: errorCode(23),
         detailedError: `Missing iss property in siop. Expected '${VerifiableCredentialConstants.TOKEN_SI_ISS}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
       };
     }
 
@@ -462,7 +476,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(24),
         detailedError: `Wrong iss property in siop. Expected '${VerifiableCredentialConstants.TOKEN_SI_ISS}'`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -472,7 +487,8 @@ export class ValidationHelpers {
         result: false,
         code: errorCode(25),
         detailedError: `Missing aud property in siop`,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidRequest,
       };
     }
 
@@ -482,7 +498,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(26),
           detailedError: `Wrong aud property in siop. Expected '${expected.audience}'`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
     }
@@ -506,7 +523,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(27),
           detailedError: `The signature on the payload in the ${(self as ValidationOptions).tokenType} is invalid`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
     } catch (err) {
@@ -516,7 +534,8 @@ export class ValidationHelpers {
         code: errorCode(28),
         detailedError: `Failed to validate signature`,
         innerError: err,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -539,52 +558,58 @@ export class ValidationHelpers {
           method: 'GET',
           headers: {
             'Content-Type': 'application/json'
-          }});
+          }
+        });
 
         if (!response.ok) {
           return {
             result: false,
-            status: 403,
             code: errorCode(29),
-            detailedError: `Could not fetch token configuration needed to validate token`
+            detailedError: `Could not fetch token configuration needed to validate token`,
+            status: this.validatorOptions.invalidTokenError,
+            wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
           };
         }
-        
+
         const config = await response.json();
         const keysUrl = config[VerifiableCredentialConstants.CONFIG_JWKS];
-        
+
         if (!keysUrl) {
           return {
             result: false,
-            status: 403,
             code: errorCode(30),
-            detailedError: `No reference to jwks found in token configuration`
+            detailedError: `No reference to jwks found in token configuration`,
+            status: this.validatorOptions.invalidTokenError,
+            wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
           };
         }
-        
+
         response = await this.validatorOptions.fetchRequest.fetch(keysUrl, 'OIDCJwks', {
           method: 'GET',
           headers: {
             'Content-Type': 'application/json'
-          }});
-          
+          }
+        });
+
         if (!response.ok) {
           return {
             result: false,
-            status: 403,
             code: errorCode(31),
-            detailedError: `Could not fetch keys needed to validate token on '${keysUrl}'`
+            detailedError: `Could not fetch keys needed to validate token on '${keysUrl}'`,
+            status: this.validatorOptions.invalidTokenError,
+            wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
           };
         }
 
         keys = await response.json();
-        
+
         if (!keys || !keys.keys) {
           return {
             result: false,
-            status: 403,
             code: errorCode(32),
-            detailedError: `No or bad jwks keys found in token configuration`
+            detailedError: `No or bad jwks keys found in token configuration`,
+            status: this.validatorOptions.invalidTokenError,
+            wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
           };
         }
 
@@ -593,9 +618,10 @@ export class ValidationHelpers {
         if (!config.issuer) {
           return {
             result: false,
-            status: 403,
             code: errorCode(33),
-            detailedError: `No issuer found in token configuration`
+            detailedError: `No issuer found in token configuration`,
+            status: this.validatorOptions.invalidTokenError,
+            wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
           };
         }
       }
@@ -603,10 +629,11 @@ export class ValidationHelpers {
       console.error(err);
       return {
         result: false,
-        status: 403,
         innerError: err,
         code: errorCode(34),
-        detailedError: `Could not fetch token configuration`
+        detailedError: `Could not fetch token configuration`,
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
 
@@ -620,12 +647,12 @@ export class ValidationHelpers {
    * @returns validationResponse.issuer The issuer found in the configuration. Should match the issuer in the token.
    */
   public async fetchKeyAndValidateSignatureOnIdToken(validationResponse: IValidationResponse, token: ClaimToken): Promise<IValidationResponse> {
-    const self: any = this ;
+    const self: any = this;
     const publicKeyResponse = await (self as IValidationOptions).fetchOpenIdTokenPublicKeysDelegate(validationResponse, token);
 
     // if we don't have a keys property, it's because we have IValidationResponse instance, no good way to check for type of an interface
-    if(!publicKeyResponse.keys) {
-      return <IValidationResponse> publicKeyResponse;
+    if (!publicKeyResponse.keys) {
+      return <IValidationResponse>publicKeyResponse;
     }
 
     const keys = publicKeyResponse.keys;
@@ -662,9 +689,10 @@ export class ValidationHelpers {
         if (!validated) {
           return {
             result: false,
-            status: 403,
             code: errorCode(35),
-            detailedError: `Could not validate token signature`
+            detailedError: `Could not validate token signature`,
+            status: this.validatorOptions.invalidTokenError,
+            wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
           };
         }
       }
@@ -678,10 +706,11 @@ export class ValidationHelpers {
       console.error(err);
       return {
         result: false,
-        status: 403,
         code: errorCode(36),
         innerError: err,
-        detailedError: `Could not validate signature on id token`
+        detailedError: `Could not validate signature on id token`,
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
   }
@@ -704,7 +733,8 @@ export class ValidationHelpers {
           result: false,
           code: errorCode(37),
           detailedError: `The presented ${(self as ValidationOptions).tokenType} is has an invalid signature`,
-          status: 403
+          status: this.validatorOptions.invalidTokenError,
+          wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
         };
       }
 
@@ -718,7 +748,8 @@ export class ValidationHelpers {
         code: errorCode(38),
         detailedError: `Failed to verify token signature`,
         innerError: err,
-        status: 403
+        status: this.validatorOptions.invalidTokenError,
+        wwwAuthenticateError: AuthenticationErrorCode.invalidToken,
       };
     }
   }