+++ lib/options/BasicValidatorOptions.ts
@@ -2,7 +2,7 @@
  *  Copyright (c) Microsoft Corporation. All rights reserved.
  *  Licensed under the MIT License. See License in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
-import { Crypto, IDidResolver, CryptoBuilder, IFetchRequest, FetchRequest, ValidationSafeguards } from '../index';
+import { Crypto, IDidResolver, CryptoBuilder, IFetchRequest, FetchRequest, ValidationSafeguards, ValidatorBuilder } from '../index';
 import IValidatorOptions from './IValidatorOptions';
 
 /**
@@ -13,11 +13,17 @@ export default class BasicValidatorOptions implements IValidatorOptions {
   private readonly _crypto: Crypto;
   private readonly _fetchRequest: IFetchRequest;
   private readonly _validationSafeguards: ValidationSafeguards;
+  private readonly _invalidTokenError: number;
 
   constructor(private _resolver?: IDidResolver) {
     this._crypto = new CryptoBuilder().build();
     this._fetchRequest = new FetchRequest();
     this._validationSafeguards = new ValidationSafeguards();
+    this._invalidTokenError = ValidatorBuilder.INVALID_TOKEN_STATUS_CODE;
+  }
+  
+  get invalidTokenError(): number {
+    return this._invalidTokenError;
   }
 
   get resolver(): IDidResolver {