+++ tests/SiopValidation.spec.ts
@@ -8,7 +8,8 @@ import { SiopValidation } from "../lib/input_validation/SiopValidation";
 import TestSetup from './TestSetup';
 import ValidationOptions from '../lib/options/ValidationOptions';
 import { IssuanceHelpers } from "./IssuanceHelpers";
-import { DidValidation, IExpectedSiop, TokenType } from "../lib/index";
+import { AuthenticationErrorCode, DidValidation, IExpectedSiop, TokenType, ValidatorBuilder } from "../lib/index";
+import VerifiableCredentialConstants from "../lib/verifiable_credential/VerifiableCredentialConstants";
 
 describe('SiopValidation', () =>
 {
@@ -39,9 +40,10 @@ describe('SiopValidation', () =>
     let siopRequest = await IssuanceHelpers.createSiopRequestWithPayload(setup, payload, siop.didJwkPrivate);
     response = await validator.validate(<string>siopRequest.rawToken);
     expect(response.result).toBeFalsy();
-    expect(response.status).toEqual(403);
+    expect(response.status).toEqual(ValidatorBuilder.INVALID_TOKEN_STATUS_CODE);
     expect(response.detailedError).toEqual(`Missing iss property in siop. Expected 'https://self-issued.me'`);
     expect(response.code).toEqual('VCSDKVaHe23');
+    expect(response.wwwAuthenticateError).toEqual(AuthenticationErrorCode.invalidRequest);
 
     // Bad iss
     payload = {
@@ -50,9 +52,10 @@ describe('SiopValidation', () =>
     siopRequest = await IssuanceHelpers.createSiopRequestWithPayload(setup, payload, siop.didJwkPrivate);
     response = await validator.validate(<string>siopRequest.rawToken);
     expect(response.result).toBeFalsy();
-    expect(response.status).toEqual(403);
+    expect(response.status).toEqual(ValidatorBuilder.INVALID_TOKEN_STATUS_CODE);
     expect(response.detailedError).toEqual(`Wrong iss property in siop. Expected 'https://self-issued.me'`);
     expect(response.code).toEqual('VCSDKVaHe24');
+    expect(response.wwwAuthenticateError).toEqual(AuthenticationErrorCode.invalidToken);
 
     // Missing aud
     payload = {
@@ -61,9 +64,10 @@ describe('SiopValidation', () =>
     siopRequest = await IssuanceHelpers.createSiopRequestWithPayload(setup, payload, siop.didJwkPrivate);
     response = await validator.validate(<string>siopRequest.rawToken);
     expect(response.result).toBeFalsy();
-    expect(response.status).toEqual(403);
+    expect(response.status).toEqual(ValidatorBuilder.INVALID_TOKEN_STATUS_CODE);
     expect(response.detailedError).toEqual(`Missing aud property in siop`);
     expect(response.code).toEqual('VCSDKVaHe25');
+    expect(response.wwwAuthenticateError).toEqual(AuthenticationErrorCode.invalidRequest);
 
     // Wrong aud
     payload = {
@@ -73,9 +77,10 @@ describe('SiopValidation', () =>
     siopRequest = await IssuanceHelpers.createSiopRequestWithPayload(setup, payload, siop.didJwkPrivate);
     response = await validator.validate(<string>siopRequest.rawToken);
     expect(response.result).toBeFalsy();
-    expect(response.status).toEqual(403);
+    expect(response.status).toEqual(ValidatorBuilder.INVALID_TOKEN_STATUS_CODE);
     expect(response.detailedError).toEqual(`Wrong aud property in siop. Expected 'https://portableidentitycards.azure-api.net/42b39d9d-0cdd-4ae0-b251-b7b39a561f91/api/portable/v1.0/card/issue'`);
     expect(response.code).toEqual('VCSDKVaHe26');
+    expect(response.wwwAuthenticateError).toEqual(AuthenticationErrorCode.invalidToken);
 
     // Bad validation
     const testValidator = new DidValidation(validationOptions, expected);